issue 52

+++ Implementacija/factpro/src/main/java/ba/unsa/etf/si2013/tim9/Ponude/PonudeTrenutnaIzmjenaForm.java
@@ -264,7 +264,7 @@ public class PonudeTrenutnaIzmjenaForm extends Shell {
           item.setText(0,Integer.toString(br_stavki+1));
           item.setText(1,u.getNaziv());
           item.setText(2, u.getOpisUsluge());
-          item.setText(3, Integer.toString(u.getCijena()));
+          item.setText(3, Double.toString(u.getCijena()));
           item.setText(4, "1");
           double cijena= u.getCijena() + u.getCijena()*0.17;
           item.setText(5,Double.toString(cijena) );
@@ -309,7 +309,7 @@ public class PonudeTrenutnaIzmjenaForm extends Shell {
         item.setText(0,Integer.toString(br_stavki+1));
         item.setText(1,k.getNaziv());
         item.setText(2, k.getOpisUsluge());
-        item.setText(3, Integer.toString(k.getCijena()));
+        item.setText(3, Double.toString(k.getCijena()));
         item.setText(4, "1");
         double cijena= k.getCijena() + k.getCijena()*0.17;
         item.setText(5,Double.toString(cijena) );
